+++ x-pack/plugins/reporting/common/constants/index.ts
@@ -26,25 +26,6 @@ import {
   PNG_REPORT_TYPE_V2,
 } from '@kbn/reporting-export-types-png-common';
 
-export const PLUGIN_ID = 'reporting';
-
-export const REPORTING_TRANSACTION_TYPE = PLUGIN_ID;
-
-export const REPORTING_SYSTEM_INDEX = '.reporting';
-
-export const JOB_COMPLETION_NOTIFICATIONS_SESSION_KEY =
-  'xpack.reporting.jobCompletionNotifications';
-
-/**
- * A way to get the client side route for the reporting redirect app.
- *
- * TODO: Add a job ID and a locator to use so that we can redirect without expecting state to
- * be injected to the page
- */
-export const getRedirectAppPath = () => {
-  return '/app/reportingRedirect';
-};
-
 export const ALLOWED_JOB_CONTENT_TYPES = [
   'application/json',
   'application/pdf',
@@ -52,8 +33,6 @@ export const ALLOWED_JOB_CONTENT_TYPES = [
   'image/png',
   'text/plain',
 ];
-
-export type JobId = string;
 
 export const reportTypes = [
   CSV_REPORT_TYPE,
@@ -88,40 +67,8 @@ export const USES_HEADLESS_JOB_TYPES = [
 
 export const DEPRECATED_JOB_TYPES = [CSV_JOB_TYPE_DEPRECATED];
 
-// Licenses
-export const LICENSE_TYPE_TRIAL = 'trial' as const;
-export const LICENSE_TYPE_BASIC = 'basic' as const;
-export const LICENSE_TYPE_CLOUD_STANDARD = 'standard' as const;
-export const LICENSE_TYPE_GOLD = 'gold' as const;
-export const LICENSE_TYPE_PLATINUM = 'platinum' as const;
-export const LICENSE_TYPE_ENTERPRISE = 'enterprise' as const;
-
-export const ILM_POLICY_NAME = 'kibana-reporting';
-
-// Usage counter types
-export const API_USAGE_COUNTER_TYPE = 'reportingApi';
-export const API_USAGE_ERROR_TYPE = 'reportingApiError';
-
-// Management UI route
-export const REPORTING_MANAGEMENT_HOME = '/app/management/insightsAndAlerting/reporting';
-
-export const REPORTING_REDIRECT_LOCATOR_STORE_KEY = '__REPORTING_REDIRECT_LOCATOR_STORE_KEY__';
-
-// Statuses
-export enum JOB_STATUSES {
-  PENDING = 'pending',
-  PROCESSING = 'processing',
-  COMPLETED = 'completed',
-  FAILED = 'failed',
-  WARNINGS = 'completed_with_warnings',
-}
-
 // Test Subjects
 export const REPORT_TABLE_ID = 'reportJobListing';
 export const REPORT_TABLE_ROW_ID = 'reportJobRow';
 
-// Job params require a `version` field as of 7.15.0. For older jobs set with
-// automation that have no version value in the job params, we assume the
-// intended version is 7.14.0
-export const UNVERSIONED_VERSION = '7.14.0';
 export * from './routes';
