+++ x-pack/plugins/reporting/public/types.ts
@@ -7,6 +7,30 @@
 
 import type { CoreSetup, CoreStart } from '@kbn/core/public';
 
+import type {
+  JobId,
+  JobStatus,
+  ReportOutput,
+  ReportSource,
+  TaskRunResult,
+} from '@kbn/reporting-common/types';
+
+/* Notifier Toasts */
+export interface JobSummary {
+  id: JobId;
+  status: JobStatus;
+  jobtype: ReportSource['jobtype'];
+  title: ReportSource['payload']['title'];
+  errorCode?: ReportOutput['error_code'];
+  maxSizeReached: TaskRunResult['max_size_reached'];
+  csvContainsFormulas: TaskRunResult['csv_contains_formulas'];
+}
+
+export interface JobSummarySet {
+  completed: JobSummary[];
+  failed: JobSummary[];
+}
+
 export interface KibanaContext {
   http: CoreSetup['http'];
   application: CoreStart['application'];
