--- x-pack/plugins/reporting/server/export_types/common/index.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export interface TimeRangeParams {
-  min?: Date | string | number | null;
-  max?: Date | string | number | null;
-}
